+++ apps/aptos/components/ConnectWalletButton.tsx
@@ -1,9 +1,10 @@
-import { Button, ButtonProps } from '@pancakeswap/uikit'
-import { WalletModalV2 } from '@pancakeswap/ui-wallets'
 import { Trans, useTranslation } from '@pancakeswap/localization'
-import { useState } from 'react'
-import { useAuth } from 'hooks/useAuth'
+import { WalletModalV2 } from '@pancakeswap/ui-wallets'
+import { Button, type ButtonProps } from '@pancakeswap/uikit'
 import { wallets } from 'config/wallets'
+import { useAuth } from 'hooks/useAuth'
+import { useState } from 'react'
+import { logGTMWalletConnectEvent } from 'utils/customGTMEventTracking'
 
 export const ConnectWalletButton = ({ children, ...props }: ButtonProps) => {
   const [open, setOpen] = useState(false)
@@ -26,6 +27,7 @@ export const ConnectWalletButton = ({ children, ...props }: ButtonProps) => {
         wallets={wallets}
         login={login}
         onDismiss={() => setOpen(false)}
+        onWalletConnectCallBack={logGTMWalletConnectEvent}
       />
     </>
   )