+++ apps/web/src/hooks/Tokens.ts
@@ -16,7 +16,8 @@ import {
   useWarningTokenList,
 } from 'state/lists/hooks'
 import { safeGetAddress } from 'utils'
-import { useToken as useToken_ } from 'wagmi'
+import { useReadContracts } from 'wagmi'
+import { erc20Abi } from 'viem'
 import useUserAddedTokens from '../state/user/hooks/useUserAddedTokens'
 import { useActiveChainId } from './useActiveChainId'
 import useNativeCurrency from './useNativeCurrency'
@@ -155,15 +156,34 @@ export function useToken(tokenAddress?: string): ERC20Token | undefined | null {
 
   const address = safeGetAddress(tokenAddress)
 
-  const token: ERC20Token | undefined = address ? tokens[address] : undefined
+  const token = address ? tokens[address] : undefined
 
-  const { data, isLoading } = useToken_({
-    address: address || undefined,
-    chainId,
+  const { data, isLoading } = useReadContracts({
+    allowFailure: false,
+    contracts: [
+      {
+        chainId,
+        address,
+        abi: erc20Abi,
+        functionName: 'decimals',
+      },
+      {
+        chainId,
+        address,
+        abi: erc20Abi,
+        functionName: 'symbol',
+      },
+      {
+        chainId,
+        address,
+        abi: erc20Abi,
+        functionName: 'name',
+      },
+    ],
     query: {
-      enabled: Boolean(!!address && !token),
+      enabled: Boolean(!token && address),
+      staleTime: Infinity,
     },
-    // consider longer stale time
   })
 
   return useMemo(() => {
@@ -172,13 +192,7 @@ export function useToken(tokenAddress?: string): ERC20Token | undefined | null {
     if (unsupportedTokens[address]) return undefined
     if (isLoading) return null
     if (data) {
-      return new ERC20Token(
-        chainId,
-        data.address,
-        data.decimals,
-        data.symbol ?? 'UNKNOWN',
-        data.name ?? 'Unknown Token',
-      )
+      return new ERC20Token(chainId, address, data[0], data[1] ?? 'UNKNOWN', data[2] ?? 'Unknown Token')
     }
     return undefined
   }, [token, chainId, address, isLoading, data, unsupportedTokens])