+++ apps/web/src/views/PositionManagers/components/TableView/ActionPanel.tsx
@@ -26,7 +26,6 @@ export const ActionPanel: React.FC<
   earningToken,
   isInCakeRewardDateRange,
   manager,
-  vaultAddress,
   managerAddress,
   managerInfoUrl,
   strategyInfoUrl,