+++ packages/pools/src/constants/pools/56.ts
@@ -14,6 +14,14 @@ export const livePools: SerializedPool[] = [
     isFinished: false,
   },
   {
+    sousId: 382,
+    stakingToken: bscTokens.cake,
+    earningToken: bscTokens.aitech,
+    contractAddress: '0x4A51749d89eB4A66b1867B5FE4C56CB04311B486',
+    poolCategory: PoolCategory.CORE,
+    tokenPerBlock: '0.055592206',
+  },
+  {
     sousId: 381,
     stakingToken: bscTokens.cake,
     earningToken: bscTokens.pepe,