+++ packages/uikit/src/components/StyledComponents/index.tsx
@@ -1,0 +1,39 @@
+import { styled } from "styled-components";
+import { Flex } from "../Box";
+import { Button } from "../Button";
+
+export const DropDownHeader = styled(Flex)`
+  width: 100%;
+  height: 40px;
+  display: flex;
+  align-items: center;
+  padding: 0px 16px;
+  box-shadow: ${({ theme }) => theme.shadows.inset};
+  border: 1px solid ${({ theme }) => theme.colors.inputSecondary};
+  border-radius: 16px;
+  background: ${({ theme }) => theme.colors.input};
+  transition: border-radius 0.15s;
+`;
+
+export const DropDownContainer = styled(Button)`
+  cursor: pointer;
+  width: 100%;
+  position: relative;
+  background: ${({ theme }) => theme.colors.input};
+  border-radius: 16px;
+  height: 40px;
+  min-width: 136px;
+  user-select: none;
+  z-index: 20;
+
+  ${({ theme }) => theme.mediaQueries.sm} {
+    min-width: 168px;
+  }
+
+  .down-icon {
+    position: absolute;
+    right: 16px;
+    top: 50%;
+    transform: translateY(-50%);
+  }
+`;