+++ packages/gauges/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @pancakeswap/gauges
 
+## 1.2.12
+
+### Patch Changes
+
+- Updated dependencies [911d107]
+  - @pancakeswap/v3-sdk@3.8.10
+
+## 1.2.11
+
+### Patch Changes
+
+- Updated dependencies [fe96bb1]
+  - @pancakeswap/v3-sdk@3.8.9
+
+## 1.2.10
+
+### Patch Changes
+
+- Updated dependencies [4628ab7]
+- Updated dependencies [8f573cf]
+  - @pancakeswap/tokens@0.6.22
+
 ## 1.2.9
 
 ### Patch Changes