+++ packages/smart-router/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @pancakeswap/smart-router
 
+## 6.0.16
+
+### Patch Changes
+
+- Updated dependencies [911d107]
+  - @pancakeswap/v3-sdk@3.8.10
+  - @pancakeswap/gauges@1.2.12
+
+## 6.0.15
+
+### Patch Changes
+
+- Updated dependencies [fe96bb1]
+  - @pancakeswap/v3-sdk@3.8.9
+  - @pancakeswap/gauges@1.2.11
+
+## 6.0.14
+
+### Patch Changes
+
+- Updated dependencies [4628ab7]
+- Updated dependencies [8f573cf]
+  - @pancakeswap/tokens@0.6.22
+  - @pancakeswap/gauges@1.2.10
+  - @pancakeswap/stable-swap-sdk@1.0.14
+
 ## 6.0.13
 
 ### Patch Changes